test(proxy): move the WEB proxy tests into the mirrored tree so the live ones carry `integration`
